tests: start implementing stage subset evaluation tests; move dummy_pipeline setup into a shared helper to avoid duplication among e2e, integration tests on dummy pipelines
